% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSAE_reg.R
\name{estSAE_reg}
\alias{estSAE_reg}
\title{based on common regression models, three type of estimators}
\usage{
estSAE_reg(area, y, x, Xbar, wts, N)
}
\arguments{
\item{area}{area id}

\item{y}{response}

\item{x}{covariates intercept will be added}

\item{Xbar}{population mean}

\item{wts}{sampling weights}

\item{N}{area size}
}
\value{
list
}
\description{
based on common regression models, three type of estimators
}
