\name{SAESLCC-package}
\alias{SAESLCC-package}
\alias{SAESLCC}
\docType{package}
\title{
\packageTitle{SAESLCC}
}
\description{
\packageDescription{SAESLCC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SAESLCC}
\packageIndices{SAESLCC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SAESLCC}

Maintainer: \packageMaintainer{SAESLCC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
