% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLCC_comp.R
\name{SLCC_comp}
\alias{SLCC_comp}
\title{this code gives estimator for comp estimator.}
\usage{
SLCC_comp(
  area,
  y,
  x,
  Xbar,
  wts,
  N,
  model,
  group = NULL,
  lambda,
  standardize = TRUE
)
}
\arguments{
\item{area}{area id}

\item{y}{response}

\item{x}{covariates, add intercept automatically}

\item{Xbar}{population mean}

\item{wts}{sampling weights}

\item{N}{popualtion size}

\item{model}{three models: "intercept" (model 1), "creg" (model 2) and "ccreg" (model 3)}

\item{group}{in model 3}

\item{lambda}{tuning vector}

\item{standardize}{TRUE or FALSE, standardize y and x or not}
}
\value{
list
}
\description{
It implies several procedures.
Step 1: use SLCC function to find group based on standardized data (using a loop). 
Step 2: use the group result to construct the new estimator
}
